+++ loop/sfm/tabpfn/__init__.py
@@ -2,11 +2,9 @@
 
 import loop.sfm.tabpfn.tradeline_long_binary as tradeline_long_binary
 import loop.sfm.tabpfn.tradeable_regressor as tradeable_regressor
-import loop.sfm.tabpfn.tradeable_regressor_manifest as tradeable_regressor_manifest
 
 
 __all__ = [
     'tradeline_long_binary',
     'tradeable_regressor',
-    'tradeable_regressor_manifest'
 ]