+++ loop/tests/test_sfm.py
@@ -22,7 +22,7 @@ def test_sfm():
         (sfm.logreg.breakout_regressor_ridge, get_klines_data_large, True, True),
         (sfm.reference.lightgbm, get_klines_data_large, True, True),
         (sfm.lightgbm.tradeable_regressor, get_klines_data_large, True, True),
-        (sfm.tabpfn.tradeable_regressor_manifest, get_klines_data_fast, True, True),
+        (sfm.tabpfn.tradeable_regressor, get_klines_data_fast, True, True),
         (sfm.lightgbm.tradeline_long_binary, get_klines_data_fast, True, False),
         (sfm.lightgbm.tradeline_multiclass, get_klines_data_fast, True, False),
         (sfm.lightgbm.tradeline_directional_conditional, get_klines_data_fast, True, False),